+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQH4564KND21VW03DJB4Y3H4"
+        "01KQHEHPRY6260PJ7RXMW89WR1"
     )
     r"""The unique identifier of the data source"""