+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQH4564Z4PJ21R7F1KWD0J05"
+        "01KQHEHPS73XH1WY3DC39117Z1"
     )
     r"""The unique identifier of the data source"""